requirements.txt
Django==3.1.4
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
requests==2.25.0
stripe==2.55.1